+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- add workaround for main app window sometimes not being shown
+
 ## [0.18.0] 2024-08-13
 
 ### Changed